x86/intel_rdt: Intel haswell Cache Allocation enumeration
Cache Allocation on hsw needs to be enumerated
separately as HSW does not have support for CPUID enumeration for Cache
Allocation. Cache Allocation is only supported on certain HSW SKUs. This
patch does a probe test for hsw CPUs by writing a CLOSid(Class of
    service id) into high 32 bits of IA32_PQR_MSR and see if the bits
stick. The probe test is only done after confirming that the CPU is HSW.
Other HSW specific quirks are:
-HSW requires the L3 cache bit mask to be at least two bits.
-Maximum CLOSids supported is always 4.
-Maximum bits support in cache bit mask is always 20.

Signed-off-by: Vikas Shivappa <vikas.shivappa@linux.intel.com>