perf/x86/intel: add --per-process option to perf-stat
This patch adds --per-process option to perf-stat. This aggregates the
counts per process and its available only when -p option is specified.
The output displays the counts for each process separately. The count
for each process is simply the sum of all the counts of the threads
belonging to the process.

Signed-off-by: Vikas Shivappa <vikas.shivappa@linux.intel.com>