perf: Optimize side-band event delivery
The perf_event_aux() function iterates all PMUs and all events in
their respective per-cpu contexts to find the events to deliver
side-band records to.

For example, the brk test case in lkp triggers many mmap operations,
which, if we're also running perf, results in many perf_event_aux()
invocations.

If we enable uncore support (even when uncore event are not used),
dozens of uncore PMUs will be iterated, which can significantly
decrease brk_test's throughput.

For example, the brk throughput without uncore PMUs is 2647573
ops_per_sec and 1768444 ops_per_sec with, a 33% reduction.

To get at the per-cpu events that need side-band records, this patch
puts these events on a per-cpu list, this avoids iterating the PMUs
and any events that do not need side-band records.

Per task events are unchanged to avoid extra overhead on the context
switch paths.

Cc: vincent.weaver@maine.edu
Cc: mingo@kernel.org
Cc: acme@redhat.com
Cc: jolsa@redhat.com
Cc: ak@linux.intel.com
Cc: tglx@linutronix.de
Cc: eranian@google.com
Cc: alexander.shishkin@linux.intel.com
Reported-by: Huang, Ying <ying.huang@linux.intel.com>
Suggested-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Signed-off-by: Kan Liang <kan.liang@intel.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Link: http://lkml.kernel.org/r/1458757477-3781-1-git-send-email-kan.liang@intel.com