x86/perf/cqm: RMID recycling algorithm fixes
The recycling algorithm did not follow a well defined policy for state
transition between the limbo, free and allocated RMIDs. It did not have
a policy to clearly separate when we just jump to stabilize vs. When we
do both recycling and stabilization.

Due to these, situations arises where RMIDs were being
taken away from monitored events even though there were enough IDs in
the limbo list. This resulted in scenarios where RMID recycling almost
made the feature unusuable in scenarios as below: Say max RMID > 150 and
we are monitoring 10 events always but other events are being monitored
periodically. What happened was that we eventually run out of RMIDs and
immediately steal the RMIDs from the 10 events that are monitored rather
than using the many RMIDs in the limbo.  Other issues included 2.
Scenarios where we only saw 2-3 RMIDs being used after a few cycles of
recycling. Lot of RMID were leaked and never being used.  3. The values
reported after recycling kicked in were far far away from being accurate
(ex: Avg b/w of 2MBps shown when actual b/w is 10MBps). This happened
due to the fact that RMIDs once recycled did not stick to the event for
a certain amount of time.

This patch handles the recycling part to deal with the above issues. It
defines policies where we do not recycle and just jump to stabilize if
there are already enough RMIDs in the limbo unless we failed stabilizing
the limbo for a long time.
Also since we fixed the scheduling already to
only have one thread, its easier to keep track of the number of IDs in
limbo and number of IDs required.

The profiling stats are used instead of looping through the entire
list of events every time to find out how many IDs are needed. The
rotation ID is removed as it was just being given to the next event that
needs an RMID. Instead we can give the event directly once we take it
out of limbo which is done in the next patch.