x86/perf/cqm: RMID recycling algorithm stabilization fixes
Some issues with the RMID recycling were 1. Scenarios where we only saw 2-3
RMIDs being used after a few cycles of recycling. Lot of RMID were
leaked and never being used.  2. The values reported after
recycling kicked in were far far away from being accurate (ex: Avg b/w
of 2MBps shown when actual b/w is 10MBps). This happened
due to the fact that RMIDs once recycled did not stick to the event for a
certain amount of time.

This is a continuation of the previous patch to fix the issues mentioned
and this time the fixes are in the stabilization part.

Removed the loop to check if the RMIDs are in limbo queue for a minimum
time every time we run the stabilization. Once they have met the minimum
time, there is no need to repeatedly check the same.
Remove the loop to check if any events require RMIDs when an ID is
available from the limbo, this is just maintained as part of the rmid
recycling profiling.

The stabilization unnecessarily used a rotation_id which was just being
allocated to the next event which needed the RMID. Removed the rotation
ID and just allocated to the event which needs RMID immediately instead
of waiting till next rmid recycling timer.

Removed the conflict check which led to RMID leaks and made it very hard
to debug how the number of RMIDs being used drastically reduced after a
few cycles of recycling.  The original purpose of conflict event code
was to manage different event types like cgroup and task, but that did
not work either as we see zero counts being reported suddenly when
conflict occurs without any error being shown. So this was code which
did not do the thing for what it was meant to do and also messed up
other parts. The handling of cgroup and task events will be handled
separately in upcoming patches.

Although the stabilization succeeded, we were still attempting to
recycle more IDs which was very confusing and leading to all kinds of
inconsistencies and also was making it cumbersome to debug.
This patch makes the stabilization policy more clear:
- We keep a stabilization target and if the target is not met, we
increase the threshold_occupancy. We decrease the threshold_occupancy if
the target is met.
- If we failed to stabilize any ID in the limbo list, we increase the
limbo_fail count which eventually makes us recycle more RMIDs later.
- At any time we only recycle 1/4 the number of RMIDs among the
max_rmid.
- The max threshold occupancy is also changed to a factor of number of
threads that can be scheduled in rather than the max RMIDs, having it
as the factor of Max rmids led to a very low value when SKUs had high
number of max RMIDs and many times took very long to stabilize the IDs.