x86/perf/cqm,mbm: MBM/cqm counting fixes
For MBM, since we report total bytes for the duration the perf counts,
we need to keep the total bytes counted every time we loose an RMID.
Introduce a rc_count variable to keep this history count in the
perf_event. Also this fixes incorrect counting when multiple perf
instances count the same PID.

Cqm event was not saving the count for all events in the group and was
saving only for the group leader which is also fixed in the patch.

If we do not keep this then we endup sending a count that may be less
than the previous count during -I perf stat option which leads to
negative numbers being reported in the perf.