x86/perf/cqm: Remove the cqm conflict code and support only task events
The conflict code was meant to handle the situations where perf calls
events with different types like cgroup and task events.

However when perf calls cgroup and task events there is inconsistent
behaviour: its randomly reporting
zero counts and rarely in an interval sometimes reports some valid
count and reports no error as well as to why the counts are zeroes. This
has reported by uses as confusing and less useful and in fact gives very
inaccurate averages. Apart from that the conflict code is leading to a
lot of inconsistencies in the recycling code and leading to a lot of
RMID leaks as seen by some of the traces by users. The leak was fixed
when the conflict code was removed along with the rmid recycling fix
patches.

The cgroup event is not supported fully because when the tasks move from
cgroup to cgroup, the cqm driver is not handling the case and the RMIDs
are still tied to the task rather than the cgroup. Hence it is giving
incorrect counting for cgroup monitoring.

This patch just removes the conflict code which is causing issues and
enables support only for task events as the other scenarios (cgroup only
    or cgroup + task event) are not handled correctly anyways.