sched: introduce the finish_arch_pre_lock_switch() scheduler hook
This hook allows architecture specific code to be called right after
perf_events' context switch but before the scheduler lock is released.

It serves two uses in this patch series (introduced in next two patches):
  1) Calls CQM's cgroup context switch code that update the current RMID
  when no perf event is active (in continuous monitoring mode).
  2) Calls __pqr_ctx_switch to perform final write to the slow
  PQR_ASSOC msr in PQR's software.

Both use cases start monitoring for the next task, a role analogous to
that of perf_event_task_sched_in.

In future series, use case (1) will be expanded by Intel's CAT update to
the next CLOSID. Since Intel's CAT is independent to perf events, the hook
to perform (1) is not suitable to be in perf, yet, must be called as close
to perf sched_in as possible.

Reviewed-by: Stephane Eranian <eranian@google.com>
Signed-off-by: David Carrillo-Cisneros <davidcc@google.com>