perf/core: introduce PERF_INACTIVE_*_READ_* flags
Some offcore and uncore events, such as the new intel_cqm/llc_occupancy/,
can be read even if the event is not active in its CPU (or in any CPU).
For package events, when the event is not running in the CPU making the
read, a freshly read value is preferable to the value stored at the last
event sched out.

There are two cases covered in this patch to allow Intel's CQM (and
potentially other per package events) to obtain freshly updated values
regardless of the scheduled event in a particular CPU. Each case is
covered by a new event::pmu_event_flag:
	1) PERF_INACTIVE_CPU_READ_PKG: An event attached to a CPU that can
	be read in any CPU in its event::cpu's package, even if inactive.
	2) PERF_INACTIVE_EV_READ_ANY_CPU: An event that can be read in any
	CPU in any package in the system even if inactive.

A consequence of reading a new value from hw on each call to
perf_event_read() is that reading and saving the event value in sched out
can be avoided since the value will never be utilized. Therefore, a PMU
that sets any of the PERF_INACTIVE_*_READ_* flags can choose not to read
in context switch, at the cost of inherit_stats not working properly.

Reviewed-by: Stephane Eranian <eranian@google.com>
Signed-off-by: David Carrillo-Cisneros <davidcc@google.com>