perf/x86/intel/cqm: encapsulate per-package RMIDs
A raw RMID value is encapsulated in a Package RMID (prmid) structure
that provides atomic updates and caches recent reads.

Also, introduce __cqm_prmid_update, a helper function that reads
occupancy of a RMID while keeping track of last_read_time to
avoid multiple reads within a short period of time when moderately
stale values are acceptable, as is the case for some scenarios in this
series (e.g. reading occupancy for limbo RMIDs, or multiple transversals
of the RMID hierarchy).

Reviewed-by: Stephane Eranian <eranian@google.com>
Signed-off-by: David Carrillo-Cisneros <davidcc@google.com>