/perf/x86/intel/cqm: Use the new perf_event_attr data for intel cqm
Intel cqm and mbm events are per package and hence are not compatible
with the way perf is aggregating the stats for --per-core and
--per-socket option. perf collects per-cpu data which is not applicable
to the cqm events we cannot collect per cpu data for cqm/mbm events.

The patch just uses the data from the perf as to what aggregate options
are enabled and sends an error for the unsupported cases rather than
reporting incorrect data.  Similarly error is returned for task events
with system-wide option are not supported.

Signed-off-by: Vikas Shivappa <vikas.shivappa@linux.intel.com>